fix: bug nas rotas privadas
